wasm-jit: data reconciliation (-reconcile)
Port `SimulationRuntime/c/dataReconciliation/dataReconciliation.cpp`
(VDI 2048 D.1, D.2 and state estimation) to the wasm-jit runtime, which
until now rejected the flag outright:

    LOG_ERROR | error | wasm-jit simulation failed:
                        -reconcile: not implemented by this runtime

The compiler half (`--preOptModules+=dataReconciliation`) already
worked; only the runtime half was missing. All 42 tests in
`openmodelica/dataReconciliation` pass with `--simCodeTarget=wasm-jit`,
and the debug logs, HTML reports and CSVs come out byte-identical to
the C runtime's — bar the run directory, the build's own version
string, and `ctime` being UTC here as everywhere else in the port.

The procedures live in `openmodelica_sim_meta::datarecon`, beside the
driver they hang off, so one copy serves the host-driven native run,
the in-wasm session driver and the wasip1 standalone command. Reports
go through `std::fs`: real files natively and under wasip1, the VFS in
the browser.

Model side: where C generates `data_function` / `setc_function` /
`setb_function` to copy variables into three `simulationInfo` arrays,
the runtime addresses the `SimData` slots directly, as `-l` already
does for its inputs and outputs. Only the symbolic Jacobians `F` and
`H` need emitted code, and they reuse the `-l` column emitter, whose
column-major output is the layout `getJacobianMatrixF` assembles by
hand. `functionDAE` (`functionLocalKnownVars` + `allEquations`) is
emitted only for a model the extraction algorithm ran on — it is not
`functionODE` + `functionAlgebraics`, which for a `when` model would
evaluate the ODE half twice.

C prints the procedure's log lines after the run's success line but
runs it while the model is still alive, so they are buffered and
replayed past the teardown hook, which is where the caller's capture
split puts post-success output.

`MODEL_FNS` was exactly 32 entries, so `present_mask` becomes a `u64`.

+++ OMCompiler/Compiler/OpenModelica.rs/openmodelica_wasm_jit/src/sim_runtime_wasmer.rs
@@ -1216,12 +1216,12 @@ pub fn build_inwasm_session(model: &SimModel) -> std::result::Result<InWasmSessi
     let table = wts(rt_inst.exports.get_table("__indirect_function_table"))?.clone();
     let n_slots = crate::model::INWASM_SLOT_NAMES.len() as u32;
     let fn_base = wts(table.grow(&mut store, n_slots, wasmer::Value::FuncRef(None)))?;
-    let mut present_mask: u32 = 0;
+    let mut present_mask: u64 = 0;
     for (slot, name) in crate::model::INWASM_SLOT_NAMES.iter().enumerate() {
         if let Ok(f) = instance.exports.get_function(name) {
             let f = f.clone();
             wts(table.set(&mut store, fn_base + slot as u32, wasmer::Value::FuncRef(Some(f))))?;
-            present_mask |= 1 << slot;
+            present_mask |= 1u64 << slot;
         }
     }
 
@@ -1250,7 +1250,7 @@ pub fn build_inwasm_session(model: &SimModel) -> std::result::Result<InWasmSessi
         return Err("CodegenWasmJit: the runtime rejected the simulation flags".to_string());
     }
 
-    let start: wasmer::TypedFunction<(u32, u32, u32, u32), i32> =
+    let start: wasmer::TypedFunction<(u32, u32, u32, u64), i32> =
         wts(rt_inst.exports.get_typed_function(&store, "rt_sim_start"))?;
     let gf = |store: &Store, name: &'static str| -> std::result::Result<wasmer::TypedFunction<(), u32>, String> {
         wts(rt_inst.exports.get_typed_function(store, name))
